+++ src/soup_backend.rs
@@ -540,7 +540,7 @@ impl SoupBackend {
             let written = match c {
                 DataType::None => rw.write_col(None::<i32>),
                 DataType::Int(i) => rw.write_col(i as i32),
-                DataType::BigInt(i) => rw.write_col(i as i64),
+                DataType::BigInt(i) => rw.write_col(i as i32),
                 DataType::Text(t) => rw.write_col(t.to_str().unwrap()),
                 dt @ DataType::TinyText(_) => rw.write_col(dt.to_string()),
                 dt @ DataType::Real(_, _) => {
